requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cloudinary==1.44.0
cryptography==45.0.3
dj-database-url==3.0.0
Django==5.1.7
django-allauth==65.9.0
django-cloudinary-storage==0.3.0
django-defender==0.9.8
django-environ==0.12.0
django-ratelimit==4.1.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.9
pycparser==2.22
redis==6.2.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
